refactor the code structure of llm_http_server
